feat: restructure header and footer components; add responsive design and dropdown functionality
