--- src/app/components/header/header.css
@@ -1,121 +1,0 @@
-.header {
-  background-color: var(--verde-oscuro);
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  z-index: 1000;
-
-
-  .container {
-    max-width: 1200px;
-    margin: 0 auto;
-    padding: 0 2rem;
-  }
-
-  .header-content {
-    margin-top: 0.5rem;
-    margin-bottom: 0.5rem;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    height: 5rem;
-  }
-
-  .logo img {
-    height: 3rem;
-    width: auto;
-    border-radius: 0.5rem;
-  }
-
-  .hamburger {
-    display: none;
-    flex-direction: column;
-    gap: 5px;
-    cursor: pointer;
-
-    span {
-      width: 25px;
-      height: 3px;
-      background-color: var(--blanco);
-      border-radius: 2px;
-    }
-  }
-
-  .nav-links {
-    list-style: none;
-    display: flex;
-    align-items: center;
-    gap: 1.5rem;
-
-    li a {
-      text-decoration: none;
-      color: var(--blanco);
-      font-weight: 600;
-      background: none;
-      border: none;
-      cursor: pointer;
-      transition: color 0.3s ease;
-    }
-
-    li a:hover {
-      color: var(--verde-turquesa);
-    }
-    }
-    .register-btn {
-      cursor: pointer;
-      background-color: var(--blanco);
-      color: var(--negro);
-      padding: 0.5rem 1.2rem;
-      font-weight: 600;
-      border: 2px solid var(--negro);
-      border-radius: 8px;
-      transition: all 0.3s ease;
-
-      &:hover {
-        background-color: var(--negro);
-        color: var(--blanco);
-      }
-  }
-
-  @media (max-width: 768px) {
-
-    .container  {
-      padding: 0 0.9rem;
-      flex-direction: row;
-      justify-content: center;
-    }
-    .hamburger {
-      display: flex;
-      position: relative; 
-      z-index: 2000;
-      margin-left: 1rem;
-    }
-
-    .nav-links {
-      position: absolute;
-      top: 100%;
-      left: 0;
-      right: 0;
-      background-color: var(--verde-oscuro);
-      flex-direction: column;
-      align-items: flex-start;
-      padding: 1rem 2rem;
-      display: none;
-      opacity: 0;
-      transform: translateY(-10px);
-      transition: all 0.3s ease-in-out;
-
-      &.active {
-        display: flex;
-        opacity: 1;
-        transform: translateY(0);
-      }
-
-      li {
-        width: 100%;
-        padding: 0.5rem 0;
-      }
-    }
-  }
-}
