--- src/app/components/header/header.ts
@@ -1,21 +1,0 @@
-import { Component } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { RouterModule } from '@angular/router';
-
-@Component({
-  selector: 'app-header',
-  imports: [CommonModule,RouterModule],
-  templateUrl: './header.html',
-  styleUrl: './header.css'
-})
-export class Header {
-  menuOpen = false;
-
-  toggleMenu() {
-    this.menuOpen = !this.menuOpen;
-  }
-
-  closeMenu() {
-    this.menuOpen = false;
-  }
-}
