+++ src/app/pages/home/home.html
@@ -1,5 +1,4 @@
-<!-- ✅ home.component.html -->
-<section class="home">
+<section class="home" id="hero">
   <div class="container">
     <div class="hero">
       <div class="text">
