+++ src/app/pages/home/home.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { RouterModule } from '@angular/router';
+import { RouterModule  } from '@angular/router';
 
 @Component({
   selector: 'app-home',
@@ -7,6 +7,6 @@ import { RouterModule } from '@angular/router';
   templateUrl: './home.html',
   styleUrl: './home.css'
 })
-export class Home {
-
+export class Home  {
+  
 }
