+++ src/app/shared/components/footer/footer.ts
